refactor(payroll): isolate dismissal loader
